Style to 1st new message screen

+++ Digital Album/Application/Gallery/GalleryViewController.h
@@ -16,7 +16,7 @@
 }
 
 @property (nonatomic, strong) DAAlbum * album;
-@property (nonatomic, assign) int startingIndex;
+@property (nonatomic, assign) NSUInteger startingIndex;
 
 -(void)toggleNavControls;
